feat: native/mobile messaging module
